Fix missing references to the 'RtuTcp' backend in Device class
Although the 'RtuTcpLayer' class was already working with the underlying
libmodbus implementation, the 'Device' class was missing some
documentation and the method 'rtu_tcp', regarding the 'RtuTcp' backend.

Since the 'RtuTcp' backend sends RTU datagrams over a TCP connection,
it would make sense to use the method 'prepareToSend' from 'RtuLayer'
and the method 'sendRawMessage' from 'TcpLayer', but we are still
considering if it is a good idea to apply multiple inheritance in this
case, since it would lead to a 'diamond problem' with both base classes
having 'NetLayer' as their common base class. In the meantime these
methods just throw an exception if called for the 'RtuTcp' backend, as a
reminder they are still not implemented.